Fix CPU-RAM OOM on the 132k corpus: float16 corpus load + free samples list
load_synthetic_data loaded the whole 132k x 2ch x 128x128 corpus as float32 (~17 GB)
AND kept the raw per-sample dict list alive (~another corpus) -> Slurm oom_kill at 48 GB.
Now store the corpus in float16 (source data is already float16; downstream casts to
float() per batch) and `del samples` after building tensors -> ~20 GB peak. train_router
casts its batch to float() (float16 corpus -> float32 CNN, was a dtype crash). Bump job
--mem 48G->64G for headroom. Verified: 2-step pretrain (experts + router) and frozen
downstream both run on the float16 corpus/eval.

+++ spectro/scripts/spectro_pretrain.py
@@ -214,7 +214,7 @@ def train_router(specs: torch.Tensor, protocol: np.ndarray, *, epochs, lr, batch
     for ep in range(epochs):
         router.train()
         for xb, yb in tr_loader:
-            xb, yb = xb.to(device), yb.to(device)
+            xb, yb = xb.to(device).float(), yb.to(device)   # corpus may be float16 -> float32 for the CNN
             opt.zero_grad()
             loss = criterion(router(_normalize_per_sample(xb)), yb)
             loss.backward(); opt.step()
@@ -222,7 +222,7 @@ def train_router(specs: torch.Tensor, protocol: np.ndarray, *, epochs, lr, batch
         correct, total = 0, 0
         with torch.no_grad():
             for xb, yb in val_loader:
-                xb = xb.to(device)
+                xb = xb.to(device).float()
                 pred = router(_normalize_per_sample(xb)).argmax(1).cpu()
                 correct += (pred == yb).sum().item(); total += yb.size(0)
         acc = correct / max(total, 1)
